requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
google-genai>=0.5.0
PyYAML>=6.0

# CLI enhancement (optional but recommended)
rich>=13.0.0
